requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
pytest==7.4.2
apache-airflow==2.6.3
six>=1.16.0
confluent-kafka
